+++ src/js/plyr.polyfilled.js
@@ -1,14 +1,13 @@
 // ==========================================================================
 // Plyr Polyfilled Build
-// plyr.js v3.2.4
+// plyr.js v3.3.10
 // https://github.com/sampotts/plyr
 // License: The MIT License (MIT)
 // ==========================================================================
 
 import 'babel-polyfill';
-
 import 'custom-event-polyfill';
-
+import 'url-polyfill';
 import Plyr from './plyr';
 
 export default Plyr;